feat: move trainer classes to another file
