+++ deeppavlov/core/trainers/__init__.py
@@ -1,0 +1,1 @@
+from .trainers import FitTrainer, NNTrainer
